timing test00

+++ test00.py
@@ -1,15 +1,18 @@
 #
 # run time test
 #
+#  mac:         3.4
+#  nectar:      22
+#  nesi build:
+#  nesi run:
 
 import time
 import numpy as np
-from multiprocessing import Process, Pipe
 
 # execution timer
 start = time.time()
 
-incr = np.float32(0.1)
+incr = np.double(0.1)
 x = incr
 for i in range(1000000):
     k = np.log(x)
